+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and **Merged pull requests**.
 The versions coincide with releases on pypi.
 
 ## [0.0.x](https://github.com/vsoch/juliart/tree/master) (0.0.x)
+ - JuliaSetAnimation class to generation animations (0.0.12)
  - adding color themes (0.0.11)
  - initial release with basic generation for random, glow, pattern colors (0.0.1)